+++ CHANGES.rst
@@ -1,5 +1,14 @@
-.. currentmodule:: flask-pydantic-spec
-VERSION 0.4.1
+.. currentmodule:: flask-pydantic-openapi
+VERSION 0.3.2
+-------------
+
+Release 2023-01-31
+
+- Add ability to add root path to config to add additional prefix in api routes
+- Add ability to hide openapi docs
+
+
+VERSION 0.4.2
 -------------
 
 Release 2022-11-14